+++ src/main/java/de/crashmash/citybuild/storage/Storage.java
@@ -44,7 +44,7 @@ public class Storage {
         this.startKickCollection = database.createCollection("StartKick")
                 .field("UUID", DataType.UUID, FieldOption.PRIMARY_KEY)
                 .field("Reason", DataType.STRING)
-                .field("Duration", DataType.LONG)
+                .field("Duration", DataType.INTEGER)
                 .field("Cooldown", DataType.LONG)
                 .create();
